tests/keyword: Move tests into `routes` submodules
