+++ src/tests/routes/mod.rs
@@ -14,6 +14,7 @@
 pub mod categories;
 pub mod category_slugs;
 pub mod crates;
+pub mod keywords;
 pub mod me;
 pub mod summary;
 pub mod versions;
